Change global logger to logrus.Entry and keep base logger
This commit changes our global logger from a generic logrus.Logger to a specific logrus.Entry, while keeping a reference to the base Logger. It provides flexibility in adding context to log entries without changing the base logger configuration. This not only adds extra details to logs but also helps in tracing and debugging.