"Add InitMiniLogger function for minimal config"
This commit introduces a new function, InitMiniLogger, which initializes a logger with minimal configuration. This function was added to provide a quicker way to set up a basic logger when extensive configuration is not necessary. It will help increase efficiency in development by reducing the time spent on logger setup.